Organization, part the first
1. Break into `main` and `lib`.
2. Add a bunch of `deny!` and `warn!` things at the top.
3. Fix all the clippy issues
4. ... Profit?
